Refactored the admin in preparation for adding another custom view.
